[Fix] Divide chat_access_token method (circular import)
