+++ chzzkpy/chat/chat_client.py
@@ -30,23 +30,25 @@ import aiohttp
 from .access_token import AccessToken
 from .enums import ChatCmd
 from .gateway import ChzzkWebSocket, ReconnectWebsocket
+from .http import ChzzkChatSession
 from .message import ChatMessage
 from .recent_chat import RecentChat
 from .state import ConnectionState
 from ..client import Client
 from ..error import LoginRequired
+from ..http import ChzzkAPISession
 
 _log = logging.getLogger(__name__)
 
 
 class ChatClient(Client):
     def __init__(
-        self,
-        channel_id: str,
-        authorization_key: Optional[str] = None,
-        session_key: Optional[str] = None,
-        chat_channel_id: Optional[str] = None,
-        loop: Optional[asyncio.AbstractEventLoop] = None,
+            self,
+            channel_id: str,
+            authorization_key: Optional[str] = None,
+            session_key: Optional[str] = None,
+            chat_channel_id: Optional[str] = None,
+            loop: Optional[asyncio.AbstractEventLoop] = None,
     ):
         super().__init__(
             loop=loop, authorization_key=authorization_key, session_key=session_key
@@ -71,6 +73,10 @@ class ChatClient(Client):
         handler = {ChatCmd.CONNECTED: self._ready.set}
         self._connection = ConnectionState(dispatch=self.dispatch, handler=handler)
         self._gateway: Optional[ChzzkWebSocket] = None
+
+    def _session_initial_set(self):
+        self._api_session = ChzzkAPISession(loop=self.loop)
+        self._game_session = ChzzkChatSession(loop=self.loop)
 
     @property
     def is_connected(self) -> bool:
@@ -142,15 +148,14 @@ class ChatClient(Client):
         await self._ready.wait()
 
     def wait_for(
-        self,
-        event: str,
-        check: Optional[Callable[..., bool]] = None,
-        timeout: Optional[float] = None,
+            self,
+            event: str,
+            check: Optional[Callable[..., bool]] = None,
+            timeout: Optional[float] = None,
     ):
         future = self.loop.create_future()
 
         if check is None:
-
             def _check(*_):
                 return True
 
@@ -163,7 +168,7 @@ class ChatClient(Client):
         return asyncio.wait_for(future, timeout=timeout)
 
     def event(
-        self, coro: Callable[..., Coroutine[Any, Any, Any]]
+            self, coro: Callable[..., Coroutine[Any, Any, Any]]
     ) -> Callable[..., Coroutine[Any, Any, Any]]:
         if not asyncio.iscoroutinefunction(coro):
             raise TypeError("function must be a coroutine.")
@@ -212,26 +217,31 @@ class ChatClient(Client):
             self._schedule_event(coroutine_function, method, *args, **kwargs)
 
     async def _run_event(
-        self,
-        coro: Callable[..., Coroutine[Any, Any, Any]],
-        *args: Any,
-        **kwargs: Any,
+            self,
+            coro: Callable[..., Coroutine[Any, Any, Any]],
+            event_name: str,
+            *args: Any,
+            **kwargs: Any,
     ) -> None:
         try:
             await coro(*args, **kwargs)
         except asyncio.CancelledError:
             pass
         except Exception as exc:
-            self.dispatch("on_error", exc, *args, **kwargs)
+            try:
+                _log.exception('Ignoring exception in %s', event_name)
+                self.dispatch("error", exc, *args, **kwargs)
+            except asyncio.CancelledError:
+                pass
 
     def _schedule_event(
-        self,
-        coro: Callable[..., Coroutine[Any, Any, Any]],
-        event_name: str,
-        *args: Any,
-        **kwargs: Any,
+            self,
+            coro: Callable[..., Coroutine[Any, Any, Any]],
+            event_name: str,
+            *args: Any,
+            **kwargs: Any,
     ) -> asyncio.Task:
-        wrapped = self._run_event(coro, *args, **kwargs)
+        wrapped = self._run_event(coro, event_name, *args, **kwargs)
         # Schedules the task
         return self.loop.create_task(wrapped, name=f"chzzk.py: {event_name}")
